+++ web/src/lib/graph/index.js
@@ -1,0 +1,4 @@
+import Graph from './Graph.vue';
+import './style.scss';
+
+export default Graph;