+++ web/src/lib/ui/components/index.js
@@ -2,7 +2,6 @@ export * from './Badge';
 export * from './Button';
 export * from './Checkbox';
 export * from './Cover';
-export * from './Graph';
 export * from './Icon';
 export * from './Menu';
 export * from './Modal';